Provide schema creation script as code
Add schema creation script generated with MySQL Workbench to make it
easier for everyone to create the schema without the GUI tool. There
are two things to consider:
  * the script includes VISIBLE keyword for indexes (due to [1]);
  * the script is manually modified to fix size of columns
    results.racing_time and results_in.racing_time sizes (due to [2]);
The schema creation script _must_ be updated manually after update
of the model in MySQL Workbench, but perhaps a better way could be
found in future.
Also make GitHub Linguist consider the new script as generated.

See #1

References:
  [1] https://bugs.mysql.com/bug.php?id=95804
  [2] https://bugs.mysql.com/bug.php?id=81123